NoMouse Option added, Melt Chest positon changed
-NoMouse Option added:
Added Checkbox to Gui, if activated it will only join the PvP but wont move or shoot - (selfplay mode with autojoin)
-Melt Chest positon changed
after melting chests game endet up in the wrong window (ranking) several times. Click position changed to avoid that. If that won't do - i have added a part that checks for that window (currently disabled)